more refactoring, removed /smart/
