+++ frontend/src/components/admin/AdminTab.vue
@@ -25,8 +25,8 @@ import Vue from 'vue';
 import {mapActions, mapState} from 'vuex';
 import {Contract} from '@/interfaces';
 import {PropType} from 'vue/types/options';
-import RoleGranter from '@/components/smart/RoleGranter.vue';
-import RoleRevoker from '@/components/smart/RoleRevoker.vue';
+import RoleGranter from '@/components/role/RoleGranter.vue';
+import RoleRevoker from '@/components/role/RoleRevoker.vue';
 import Hint from '@/components/Hint.vue';
 import QuestsAdmin from './AdminTabs/QuestsAdmin.vue';
 import CBKLandAdmin from './AdminTabs/CBKLandAdmin.vue';
