+++ frontend/src/components/balance/DustBalanceDisplay.vue
