+++ frontend/src/components/character/CharacterDisplay.vue
@@ -121,7 +121,7 @@ interface StoreMappedActions {
   getCharacterBusyStatus(payload: { characterId: string | number }): Promise<number>;
 }
 import i18n from '@/i18n';
-import EarningsCalculator from '../smart/EarningsCalculator.vue';
+import EarningsCalculator from '../earningsCalculator/EarningsCalculator.vue';
 
 export default Vue.extend({
   props: ['toggled', 'currentPath'],
