+++ frontend/src/components/earningsCalculator/EarningsCalculator.vue
