+++ frontend/src/components/loader/Loader.vue
