+++ frontend/src/components/nftList/NftList.vue
