+++ frontend/src/components/quests/QuestSubmissionModal.vue
@@ -63,9 +63,9 @@
 import Vue from 'vue';
 import WeaponGrid from '@/components/layout/WeaponGrid.vue';
 import {mapActions, mapGetters, mapState} from 'vuex';
-import NftList, {NftIdType} from '@/components/smart/NftList.vue';
+import NftList, {NftIdType} from '@/components/nftList/NftList.vue';
 import {Quest, QuestTemplateType, Rarity, RequirementType, RewardType} from '@/views/Quests.vue';
-import DustBalanceDisplay from '@/components/smart/DustBalanceDisplay.vue';
+import DustBalanceDisplay from '@/components/balance/DustBalanceDisplay.vue';
 import NftIcon from '@/components/NftIcon.vue';
 import {PropType} from 'vue/types/options';
 import {Nft} from '../../interfaces/Nft';
