+++ frontend/src/components/role/RoleRevoker.vue
