+++ frontend/src/views/Blacksmith.vue
@@ -620,11 +620,11 @@ import Vue from 'vue';
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex';
 import WeaponIcon from '../components/WeaponIcon.vue';
 import { BModal } from 'bootstrap-vue';
-import NftList from '@/components/smart/NftList.vue';
+import NftList from '@/components/nftList/NftList.vue';
 import { Contracts, IState } from '@/interfaces';
 import { ISpecialWeaponsManagerState } from '@/store/specialWeaponsManager';
 import { Accessors } from 'vue/types/options';
-import DustBalanceDisplay from '@/components/smart/DustBalanceDisplay.vue';
+import DustBalanceDisplay from '@/components/balance/DustBalanceDisplay.vue';
 import { fromWeiEther, toBN } from '@/utils/common';
 import i18n from '@/i18n';
 import Events from '../events';
