+++ frontend/src/views/Bridge.vue
@@ -301,11 +301,11 @@ import {mapActions, mapGetters, mapState} from 'vuex';
 import {isNftType, Nft, NftTransfer, TransferedNft} from '@/interfaces/Nft';
 import {Accessors} from 'vue/types/options';
 import {Contract, Contracts, IState} from '@/interfaces';
-import {NftIdType} from '@/components/smart/NftList.vue';
+import {NftIdType} from '@/components/nftList/NftList.vue';
 import {Characters, Shields, Weapons} from '../../../build/abi-interfaces';
 import WeaponGrid from '../components/layout/WeaponGrid.vue';
 import CharacterList from '../components/character/CharacterList.vue';
-import NftList from '../components/smart/NftList.vue';
+import NftList from '../components/nftList/NftList.vue';
 import CurrencyConverter from '../components/CurrencyConverter.vue';
 import {fromWeiEther, toBN} from '@/utils/common';
 import config from '../../app-config.json';
