+++ frontend/src/views/NftDisplay.vue
@@ -102,11 +102,11 @@ import { mapActions, mapGetters } from 'vuex';
 import { isNftType, Nft, allNftTypes } from '../interfaces/Nft';
 import { Accessors } from 'vue/types/options';
 import { Contract, Contracts } from '@/interfaces';
-import { NftIdType } from '@/components/smart/NftList.vue';
+import { NftIdType } from '@/components/nftList/NftList.vue';
 import { Characters, Shields, Weapons } from '../../../build/abi-interfaces';
 import WeaponGrid from '../components/layout/WeaponGrid.vue';
 import CharacterList from '../components/character/CharacterList.vue';
-import NftList from '../components/smart/NftList.vue';
+import NftList from '../components/nftList/NftList.vue';
 import { fromWeiEther } from '@/utils/common';
 import BigNumber from 'bignumber.js';
 import CurrencyConverter from '../components/CurrencyConverter.vue';
